requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
appdirs==1.4.4            # via black
attrs==19.3.0             # via black, pytest
black==19.10b0            # via -r requirements/dev.in
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via -r requirements/base.in, black, pip-tools
coverage==5.2.1           # via pytest-cov
idna==2.10                # via requests
iniconfig==1.0.1          # via pytest
isort==5.2.2              # via -r requirements/dev.in
kafka-python==2.0.1       # via -r requirements/base.in
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
pathspec==0.8.0           # via black
pip-tools==5.3.1          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.5    # via -r requirements/base.in
py==1.9.0                 # via pytest
pydantic==1.6.1           # via -r requirements/base.in
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements/dev.in
pytest==6.0.1             # via -r requirements/dev.in, pytest-cov
regex==2020.7.14          # via black
requests-mock==1.8.0      # via -r requirements/dev.in
requests==2.24.0          # via -r requirements/base.in, requests-mock
six==1.15.0               # via packaging, pip-tools, requests-mock
toml==0.10.1              # via black, pytest
typed-ast==1.4.1          # via black
urllib3==1.25.10          # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
